requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
pandas==1.4.2
SQLAlchemy==1.4.35
pysnowflake==0.1.3
flask_RESTful==0.3.9
snowflake-sqlalchemy==1.3.3
